agent/requirements.txt
duckduckgo-search
requests
html2text
mecab-python3
PyYAML
unidic
pydantic